requirements.txt
Flask
flask_mysql